+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.122.0] - 2020-06-25
+
 ### Fixed
 
 - `FilterBar` clear icon disabled style.